Fix header search icon scaling
The css reset used in this project restricts images' max width to 100%,
to stop them overflowing their parent container. However, the grid
column for the search icon had its width set to 'auto' meaning when
space ran out, the column would scale down to 0px (and thus the icon's
width would also be set to 0px).

Setting the column width to 'max-content' sets the column to be exactly
the icon's size, prevents scaling down, and resolves this issue.