Make card heights consistent with each other
Use 'fr' instead of 'min-content' so all card rows are the same height.
This does mean if a card has a particularly long description, all card
rows assume the height needed to show it.

Ideally titles and descriptions should be truncated to some number
of lines, but given this project's small scope and the time I have
already spent on it, it is not something I will implement now.